api-constantec/requirements_dev.txt
fastapi
uvicorn[standard]
black
isort
sqlalchemy
Schemas
pydantic
typing
datetime
pyodbc
passlib[bcrypt]
bcrypt<4.0
python-jose[cryptography]
factory_boy
ipython
sqladmin
itsdangerous==2.2.0